flask_app/requirements.txt
python-dotenv==1.0.1
pyjwt==2.9.0
bcrypt==4.2.0
Flask==2.2.5
Werkzeug==2.2.3
mysql-connector-python==8.0.28